Removing alt provider in favor of configProvider
