+++ .gitignore
@@ -39,3 +39,5 @@ next-env.d.ts
 *.env.(local|production)
  
 *.code-workspace
+
+demo/
