+++ src/app/reviews/[slug]/review.tsx
@@ -1,13 +1,21 @@
+"use client"
+
+
+import { useContext } from "react";
 import Link from "next/link";
 import Text from "@/components/text/text";
 import Icon from "@/components/icons/icon";
 import { Company } from "../columns";
+import { Config, ConfigContext, getAltString } from "@/lib/configProvider";
 
 const adjustedReviewDisclaimerString = "Note: The Adjusted Review Count and Rating reflect only reviews with both text and a rating.";
 
-export function Review({ data, altObj }: Readonly<{ data: Company; altObj: { [key: string]: string } }>) {
+
+export function Review({ data }: Readonly<{ data: Company; }>) {
   // Note to self: this is to display the adjusted data, the disclaimer object is different. E.g. Son-Rise -> PURE
   const hasAdjustedReviewValue: boolean = data.review_count != data.adjusted_review_count;
+  const { alt }: Config = useContext(ConfigContext)
+  const altObj: { [key: string]: string } = ["review_count", "average_rating", "adjusted_review_count", "adjusted_average_rating"].reduce((obj, curr) => ({ ...obj, [curr]: getAltString(alt, curr, data[curr]) }), {});
 
   return (
     <>
